moved all common code under /utils
