+++ ExtractUserDataActivity/handler.ts
@@ -45,10 +45,10 @@ import {
 import { iteratorToArray } from "io-functions-commons/dist/src/utils/documentdb";
 import { readableReport } from "italia-ts-commons/lib/reporters";
 import { FiscalCode, NonEmptyString } from "italia-ts-commons/lib/strings";
+import { NotificationModel } from "../utils/models/notification";
 import { generateStrongPassword, StrongPassword } from "../utils/random";
 import { AllUserData, MessageContentWithId } from "../utils/userData";
 import { getEncryptedZipStream } from "../utils/zip";
-import { NotificationModel } from "./notification";
 
 import * as yaml from "yaml";
