+++ utils/models/message.ts
@@ -2,7 +2,7 @@ import * as DocumentDb from "documentdb";
 import { Either } from "fp-ts/lib/Either";
 import { MessageModel as MessageModelBase } from "io-functions-commons/dist/src/models/message";
 import { FiscalCode, NonEmptyString } from "italia-ts-commons/lib/strings";
-import * as DocumentDbUtils from "../utils/documentdb";
+import * as DocumentDbUtils from "../documentdb";
 
 export class MessageModel extends MessageModelBase {
   public async deleteMessage(
