+++ utils/models/message_status.ts
@@ -1,24 +1,24 @@
 import * as DocumentDb from "documentdb";
 import { Either } from "fp-ts/lib/Either";
 import {
-  NOTIFICATION_STATUS_MODEL_ID_FIELD,
-  NOTIFICATION_STATUS_MODEL_PK_FIELD,
-  NotificationStatusModel as NotificationStatusModelBase,
-  RetrievedNotificationStatus
-} from "io-functions-commons/dist/src/models/notification_status";
+  MESSAGE_STATUS_MODEL_ID_FIELD,
+  MESSAGE_STATUS_MODEL_PK_FIELD,
+  MessageStatusModel as MessageStatusModelBase,
+  RetrievedMessageStatus
+} from "io-functions-commons/dist/src/models/message_status";
 import { NonEmptyString } from "italia-ts-commons/lib/strings";
-import * as DocumentDbUtils from "../utils/documentdb";
+import * as DocumentDbUtils from "../documentdb";
 
-export class NotificationStatusModel extends NotificationStatusModelBase {
-  public async deleteNotificationStatusVersion(
-    notificationId: NonEmptyString,
+export class MessageStatusModel extends MessageStatusModelBase {
+  public async deleteMessageStatusVersion(
+    messageId: NonEmptyString,
     documentId: NonEmptyString
   ): Promise<Either<DocumentDb.QueryError, string>> {
     return DocumentDbUtils.deleteDocument(
       this.dbClient,
       this.collectionUri,
       documentId,
-      notificationId
+      messageId
     );
   }
 
@@ -28,13 +28,13 @@ export class NotificationStatusModel extends NotificationStatusModelBase {
    */
   public findAllVersionsByModelId(
     modelId: NonEmptyString
-  ): DocumentDbUtils.IResultIterator<RetrievedNotificationStatus> {
+  ): DocumentDbUtils.IResultIterator<RetrievedMessageStatus> {
     return DocumentDbUtils.findAllVersionsByModelId(
       this.dbClient,
       this.collectionUri,
-      NOTIFICATION_STATUS_MODEL_ID_FIELD,
+      MESSAGE_STATUS_MODEL_ID_FIELD,
       modelId,
-      NOTIFICATION_STATUS_MODEL_PK_FIELD,
+      MESSAGE_STATUS_MODEL_PK_FIELD,
       modelId
     );
   }
